+++ src/main/java/examplemod/examples/items/ExampleBarItem.java
@@ -3,9 +3,9 @@ package examplemod.examples.items;
 import necesse.inventory.item.Item;
 import necesse.inventory.item.matItem.MatItem;
 
-public class ExampleOreItem extends MatItem {
+public class ExampleBarItem extends MatItem {
 
-    public ExampleOreItem() {
+    public ExampleBarItem() {
         super(500, Item.Rarity.UNCOMMON);
 
     }
